Rename local client to fs client
